requirements.txt
docker==6.1.3
GitPython==3.1.31
gunicorn==21.2.0
Flask==3.0.0
Flask-HTTPAuth==4.8.0
Werkzeug==3.0.1
DateTime==5.2